refactor(deploy): hoist the step-7 gate policy into a shared kernel module
The policy was declared host-neutral yet copied ~20 lines (comment
included) into both runners — drift-prone by construction and against
deploy/'s layout rule (neutral kernel at top, hosts vertical). Extract
deploy/registration-gate.ts: track outcomes + manual-notice surfacing +
the gate message; only the per-host retry remediation is a parameter.
